Move storage tests over to store.rs
